Fix if else parser, closes #135
Issue #133 introduced a leftover in the code that
has gone unnoticed since our tests are very brief:
long-story short, if you had any working code after
and if-else, the  parser would skip a token, resulting
in *fun times*.

This removes the leftover code that was there by mistake.
Preparing to rollout `1.0.1` as this is kind of a big deal.