Go modules
Last week @mingwho introduced me to `go mod` and,
even though this is the 20th attempt at implementing
vendoring/modules in Golang, it looks like this one might
be it.

Mandatory XKCD:

![http://imgs.xkcd.com/comics/standards.png](http://imgs.xkcd.com/comics/standards.png)

:)

This PR converts the `go get` commands in the dockerfile
with a simple `go mod vendor` which downloads all dependencies
locally, in the `vendor` folder.

Added mod and sum files as well to complete the deal.