hash.str() should convert to a valid JSON string, closes #197
The main issue here is that when we convert to a string, we use the
raw value of the AST object. The raw value for a string is `string`,
but that is not valid json (should be `"string"`).

Therefore, I added a new Json() method on all AST objects that
should return a compatible JSON string, and it's used when we
want to convert an hash (and all its key-values) to strings.

```
⧐  {"string": "value", "array": [1,2,3,{"hash": [1,2,null]}]}.str()

{"array": [1, 2, 3, {"hash": [1, 2, null]}], "string": "value"}
```