Allow negative in expressions, closes #239
In expressions used to return their own object.Boolean.
When we use the bang operator against them we check if
result = object.TRUE or object.FALSE. Since they returned
their own object.Boolean, which is not equal to object.TRUE|FALSE,
negation would fail (always returning false).

With this change I actually got rid of all the checks we do
when evaluating a bang expression and simply rely on the logic
behind `isTruthy`, which will avoid duplication and incoherence
in the future.