Deprecate / "unfavour" the legacy command syntax, closes #220
The `$()` syntax for commands is cool and works ok,
but presents issues with parsing and escaping. With
this PR this syntax has been removed from most of the
docs, replaced by the simpler backtick commands.

Since `$()` is still useful for embedding commands,
we kept a section of the docs dedicated to explaining
when you should use this vs backticks.