Welcome to the ABS playground!
We have created a new abs distribution that compiles
to WASM. This distribution can then be loaded on the main ABS
website to run ABS code directly in the browser.

In order to do so, I have updated all print functions to use
an io.Writer rather than the stdout -- as in the JS environment
the standard output is the console, preventing you to actually
redirect the output elsewhere.

By default, the REPL will use os.Stdout as io.Writer.