Reverse search through Ctrl+R, closes #206
This PR adds the ability to do reverse search in the ABS shell.
It works a little bit differently than the regular bash reverse search
as it's not multi-line.

Start typing the string you want to search for, then type ctrl + R:
Abs will show you the closes command that matches that search in the
history. At every Ctrl + R, ABS will go back and look for an older
matching command in the history.