source vs require, closes #272
Now `require` is not an alias of `source` aymore.

`source`, like in bash, imports a script and gives it full
access to the global environment.

`require` instead imports the script, but does not allow it
to read or set variables at the global level. If a required
scripts needs to "pass" variables to the caller, it can simply
`return` them and they will be available to th caller.

Example -- say this is our module, `ip-finder.abs`:

```
return f() {
  return `curl icanhazip.com`
}
```

Then we can simply use it like this:

```
ip_finder = require("ip-finder.abs")

echo("My IP is %s", ip_finder())
```