Optional chaining, closes #296
This PR adds optional chaining to the language, allowing to
access "deep" properties and methods without having to validate
the whole object hierarchy:

```
a.b.c.d.e()
```

would return an error if any step of the way is `null`.

This can be rewritten as:

```
a?.b?.c?.d?.e()
```

and would return `null`, without breaking a script.

Largely inspired by [this](https://chromestatus.com/feature/5668249494618112).