Use a more flexible structure for the spec argument to validateVarArgs (#309)
* Use a more flexible structure for the spec argument to validateVarArgs

validateVarArgs' argument specification now uses a format that supports
multiple argument signatures, each of whose arguments are in turn a list
of supported types. For example, to achieve:

  foo(string, [string|number])
  foo(string)

Your argument would be:

  {
    {{STRING_OBJ}, {STRING_OBJ, NUMBER_OBJ}},
    {{STRING_OBJ}},
  }

Matching is done in order, giving priority to specs by lower index.

Additionally, validateVarArgs is now smart enough to figure out the
minimum and maximum number of arguments on its own.

When validateVarArgs fails to match a spec, it will return an error with
a usage string showing the alternate calling signatures. On success, it
will return the index of the matched signature as the second return
value.

* arr.join() now defaults to an empty string as glue when no argument is specfied

* validateVarArgs' error message should begin with 'Usage:'

* Remove debug line

* split() defaults to a single space

* Update docs for join(), split()