Index / hash assignments gotchas, closes #315
They were not properly handling the case where the final
character is a `;`, when they should "consume" it and move
forward. Instead, the parser would parse the assignment and
move onto the next block, which was a single `;`, causing it
to fail.

A better fix for this could be allowing `;` as valid blocks
that are simply ignore, which would also solve expressions
such as `x = 1;;;`, which are now erroring out.