Fixed mixing property and index expression assignment
The way we assign to properties or indexes is pretty
awkward as we temporarely store in the parser the
last parsed expression -- with the caveat that we store
2 of these values, one for properties and one for indexes.
When either of them gets set, we need to "consume" the other one.

This probably needs to be rethought at some point but for
the time being this is a simple fix.