Imported modules should be cached / a singleton, closes #427
With this change, a module that's been imported is now cached.
This means that every time you call `require('module')` you will
receive back the same evaluated result, meaning changes you make
on that result will be persisted. This only applies to `require`
and not `source`, as `source` is intended to be used in order to
always evaluate a block of fresh code in the current environment.

Before:

```
$ require('@runtime').name = "xxx"
$ require('@runtime').name
abs
```

After:

```
$ require('@runtime').name = "xxx"
$ require('@runtime').name
xxx
```